News_api/requirements.txt
django
djangorestframework==3.13.1
psycopg2